GraphQL schema completion and Query fixes
GraphQL schema tested and working, but due to some Appsync issues, composite keys do not work well with @belongsTo (and most likely @hasMany).  This is not a big issue as the queries that this wouldve helped with is not very common, and still will benefit for indexed keys.

Current Task
 - Continue to convert queries
 - Fix SiteManager site